--- .dockerignore
@@ -1,4 +1,0 @@
-Dockerfile
-.dockerignore
-node_modules
-npm-debug.log
