+++ Assignment02/.dockerignore
@@ -1,0 +1,4 @@
+Dockerfile
+.dockerignore
+node_modules
+npm-debug.log
