+++ Assignment02/config/kind.config.yaml
@@ -1,0 +1,11 @@
+# Save this to 'kind.config.yaml'
+
+kind: Cluster
+apiVersion: kind.sigs.k8s.io/v1alpha3
+nodes:
+  - role: control-plane
+    extraPortMappings:
+      - containerPort: 30080
+        hostPort: 80
+        listenAddress: "0.0.0.0"
+        protocol: TCP
