+++ Assignment02/manifests/2_helloworld_deploy_ports.yaml
@@ -1,0 +1,22 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  labels:
+    app: example1
+  name: example1
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: example1
+  template:
+    metadata:
+      labels:
+        app: example1
+    spec:
+      containers:
+      - image: nginx:latest
+        name: nginx
+        ports:
+        - containerPort: 80
+          name: nginx
