+++ Assignment02/manifests/3_helloworld_service.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: example1
+  labels:
+    app: example1
+spec:
+  type: NodePort
+  selector:
+    app: example1
+  ports:
+    - protocol: TCP
+      targetPort: 80
+      port: 80
+      nodePort: 30080
