+++ Assignment02/manifests/4_helloworld_deploy_content.yaml
@@ -1,0 +1,33 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  labels:
+    app: example1
+  name: example1
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: example1
+  template:
+    metadata:
+      labels:
+        app: example1
+    spec:
+      containers:
+      - image: nginx:latest
+        name: nginx
+        ports:
+        - containerPort: 80
+          name: nginx
+        volumeMounts:
+        - name: htmlcontent
+          mountPath: "/usr/share/nginx/html/"
+          readOnly: true
+      volumes:
+      - name: htmlcontent
+        configMap:
+          name: index.html
+          items:
+          - key: index.html
+            path: index.html
