+++ Assignment02/manifests/6_service_using_local_registry.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: example2
+  labels:
+    app: example2
+spec:
+  type: NodePort
+  selector:
+    app: example2
+  ports:
+    - protocol: TCP
+      targetPort: 8080
+      port: 80
+      nodePort: 30080
