+++ Assignment02/scripts/build-aws.sh
