--- package.json
@@ -1,18 +1,0 @@
-{
-  "name": "helloworld",
-  "description": "Simple hello world sample in Node",
-  "version": "1.0.0",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js"
-  },
-  "engines": {
-    "node": ">= 12.0.0"
-  },
-  "author": "Google LLC",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "express": "^4.17.1"
-  }
-}
